+++ epoch2/cuda/ee_mumu/SubProcesses/P1_Sigma_sm_epem_mupmum/check_sa.cc
@@ -13,7 +13,7 @@
 #include "mgOnGpuTypes.h"
 
 #ifdef __CUDACC__
-#include "rambo.cc"
+#include "grambo.cu"
 #else
 #include "rambo.h"
 #endif
@@ -288,7 +288,7 @@ int main(int argc, char **argv)
     std::vector<double> commonRnd = commonRandomPromises[iiter].get_future().get();
     assert( nRnarray == static_cast<int>( commonRnd.size() ) );
     // NB (PR #45): memcpy is strictly needed only in CUDA (copy to pinned memory), but keep it also in C++ for consistency
-    memcpy( hstRnarray.get(), commonRnd.data(), nRnarray * sizeof(hstRnarray[0]) );
+    memcpy( hstRnarray, commonRandomNumbers.data(), nbytesRnarray );
 #elif defined __CUDACC__
 #ifdef MGONGPU_CURAND_ONDEVICE
     grambo2toNm0::generateRnarray( rnGen, devRnarray.get(), nevt );
@@ -714,33 +714,50 @@ int main(int argc, char **argv)
 #endif
              << "\"Curand generation\": "
 #ifdef __CUDACC__
-#if defined MGONGPU_CURAND_ONDEVICE
-             << "\"DEVICE (CUDA code)\"," << std::endl
+#if defined MGONGPU_COMMONRAND_ONHOST
+             << "\"COMMON RANDOM HOST (CUDA code)\"," << std::endl
+#elif defined MGONGPU_CURAND_ONDEVICE
+             << "\"CURAND DEVICE (CUDA code)\"," << std::endl
 #elif defined MGONGPU_CURAND_ONHOST
-             << "\"HOST (CUDA code)\"," << std::endl
+             << "\"CURAND HOST (CUDA code)\"," << std::endl
 #endif
 #else
-             << "\"HOST (C++ code)\"," << std::endl
+#if defined MGONGPU_COMMONRAND_ONHOST
+             << "\"COMMON RANDOM (C++ code)\"," << std::endl
+#else
+             << "\"CURAND (C++ code)\"," << std::endl
+#endif
 #endif
              << "\"NumberOfEntries\": " << niter << "," << std::endl
       //<< std::scientific // Not sure about this
-             << "\"TotalTimeInWaveFuncs\": "
-             << "\"" << std::to_string(sumwtim) << " sec\"," << std::endl
-             << "\"MeanTimeInWaveFuncs\": "
-             << "\"" << std::to_string(meanwtim) << " sec\"," << std::endl
-             << "\"StdDevTimeInWaveFuncs\": "
-             << "\"" << std::to_string(stdwtim) << " sec\"," << std::endl
-             << "\"MinTimeInWaveFuncs\": "
-             << "\"" << std::to_string(minwtim) << " sec\"," << std::endl
-             << "\"MaxTimeInWaveFuncs\": "
-             << "\"" << std::to_string(maxwtim) << " sec\"," << std::endl
+             << "\"TotalTime[Rnd+Rmb+ME] (123)\": \""
+             << std::to_string(sumgtim+sumrtim+sumwtim) << " sec\","
+             << std::endl
+             << "\"TotalTime[Rambo+ME] (23)\": \""
+             << std::to_string(sumrtim+sumwtim) << " sec\"," << std::endl
+             << "\"TotalTime[RndNumGen] (1)\": \""
+             << std::to_string(sumgtim) << " sec\"," << std::endl
+             << "\"TotalTime[Rambo] (2)\": \""
+             << std::to_string(sumrtim) << " sec\"," << std::endl
+             << "\"TotalTime[MatrixElems] (3)\": \""
+             << std::to_string(sumwtim) << " sec\"," << std::endl
+             << "\"MeanTimeInMatrixElems\": \""
+             << std::to_string(meanwtim) << " sec\"," << std::endl
+             << "\"MinTimeInMatrixElems\": \""
+             << std::to_string(minwtim) << " sec\"," << std::endl
+             << "\"MaxTimeInMatrixElems\": \""
+             << std::to_string(maxwtim) << " sec\"," << std::endl
       //<< "ProcessID:                = " << getpid() << std::endl
       //<< "NProcesses                = " << process.nprocesses << std::endl
              << "\"TotalEventsComputed\": " << nevtALL << "," << std::endl
-             << "\"RamboEventsPerSec\": "
-             << "\"" << std::to_string(nevtALL/sumrtim) << " sec^-1\"," << std::endl
-             << "\"MatrixElemEventsPerSec\": "
-             << "\"" << std::to_string(nevtALL/sumwtim) << " sec^-1\"," << std::endl
+             << "\"EvtsPerSec[Rnd+Rmb+ME](123)\": \""
+             << std::to_string(nevtALL/(sumgtim+sumrtim+sumwtim))
+             << " sec^-1\"," << std::endl
+             << "\"EvtsPerSec[Rmb+ME] (23)\": \""
+             << std::to_string(nevtALL/(sumrtim+sumwtim)) << " sec^-1\","
+             << std::endl
+             << "\"EvtsPerSec[MatrixElems] (3)\": \""
+             << std::to_string(nevtALL/sumwtim) << " sec^-1\"," << std::endl
              << "\"NumMatrixElements(notNan)\": " << nevtALL - nnan << "," << std::endl
              << std::scientific
              << "\"MeanMatrixElemValue\": "
